+++ src/chapter_01_math/problem_001_sum_of_naturals_divisible_by_3_or_5.cpp
@@ -25,7 +25,7 @@ namespace tmcppc::problem_1 {
 
     std::vector<size_t> naturals_divisible_by_3_or_5_up_to_limit_v2(size_t limit) {
         return ranges::views::iota(0)
-            | ranges::views::take(limit + 1)
+            | ranges::views::take(static_cast<long>(limit + 1))
             | ranges::views::filter([](auto i) { return (i % 3 == 0) or (i % 5 == 0); })
             | ranges::to<std::vector<size_t>>();
     }
@@ -38,7 +38,7 @@ void problem_1_main(std::istream& is, std::ostream& os) {
     auto n{ rtc::console::read_positive_number(is, os, "Please enter a number (>= 0): ", 0) };
 
     for (auto f : { naturals_divisible_by_3_or_5_up_to_limit_v1 , naturals_divisible_by_3_or_5_up_to_limit_v2 }) {
-        auto v{ f(n) };
+        auto v{ f(static_cast<size_t>(n)) };
 
         fmt::print(os, "The sum of all natural numbers divisible by either 3 or 5 and up to {} is:\n\t{} {}\n\n",
             n, std::accumulate(std::cbegin(v), std::cend(v), static_cast<size_t>(0)), v);
