+++ src/chapter_01_math/problem_006_abundant_numbers.cpp
@@ -30,7 +30,7 @@ void problem_6_main(std::istream& is, std::ostream& os) {
     auto limit{ rtc::console::read_positive_number(is, os, "Please enter a number (>= 1): ", 1) };
 
     fmt::print(os, "Abundant numbers up to {} [list of divisors] (and their abundance):\n", limit);
-    const auto& result{ abundant_numbers_up_to(limit) };
+    const auto& result{ abundant_numbers_up_to(static_cast<size_t>(limit)) };
     if (result.empty()) {
         fmt::print(os, "\t[]\n");
     } else {
